+++ src/js/popup/reduxs/bookmark/saga/utils/converters.test.js
@@ -116,7 +116,7 @@ describe('simulateBookmark', () => {
       isUnmodifiable: true,
       parentId: '',
       storageIndex: -1,
-      title: '',
+      title: 'ka',
       type,
       url: ''
     })